+++ src/condition.h
@@ -76,7 +76,7 @@ public:
 	virtual bool executeCondition(Creature* creature, int32_t interval);
 	virtual void endCondition(Creature* creature) = 0;
 	virtual void addCondition(Creature* creature, const Condition* condition) = 0;
-	virtual uint32_t getIcons() const;
+	virtual uint16_t getIcons() const;
 	ConditionId_t getId() const { return id; }
 	uint32_t getSubId() const { return subId; }
 
@@ -128,7 +128,7 @@ public:
 	bool executeCondition(Creature* creature, int32_t interval) override;
 	void endCondition(Creature* creature) override;
 	void addCondition(Creature* creature, const Condition* condition) override;
-	uint32_t getIcons() const override;
+	uint16_t getIcons() const override;
 
 	ConditionGeneric* clone() const override { return new ConditionGeneric(*this); }
 };
@@ -256,7 +256,7 @@ public:
 	bool executeCondition(Creature* creature, int32_t interval) override;
 	void endCondition(Creature* creature) override;
 	void addCondition(Creature* creature, const Condition* condition) override;
-	uint32_t getIcons() const override;
+	uint16_t getIcons() const override;
 
 	ConditionDamage* clone() const override { return new ConditionDamage(*this); }
 
@@ -309,7 +309,7 @@ public:
 	bool executeCondition(Creature* creature, int32_t interval) override;
 	void endCondition(Creature* creature) override;
 	void addCondition(Creature* creature, const Condition* condition) override;
-	uint32_t getIcons() const override;
+	uint16_t getIcons() const override;
 
 	ConditionSpeed* clone() const override { return new ConditionSpeed(*this); }
 
@@ -425,7 +425,7 @@ public:
 		}
 	}
 
-	uint32_t getIcons() const override;
+	uint16_t getIcons() const override;
 	void endCondition(Creature* creature) override;
 	bool startCondition(Creature* creature) override;
 	bool setParam(ConditionParam_t param, int32_t value) override;
@@ -450,7 +450,7 @@ public:
 	bool startCondition(Creature* creature) override;
 	void endCondition(Creature* creature) override;
 	void addCondition(Creature* creature, const Condition* addCondition) override;
-	uint32_t getIcons() const override;
+	uint16_t getIcons() const override;
 
 	bool setParam(ConditionParam_t param, int32_t value) override;
