+++ src/protocolgame.h
@@ -136,7 +136,7 @@ private:
 	void sendOpenPrivateChannel(const std::string& receiver);
 	void sendToChannel(const Creature* creature, SpeakClasses type, const std::string& text, uint16_t channelId);
 	void sendPrivateMessage(const Player* speaker, SpeakClasses type, const std::string& text);
-	void sendIcons(uint32_t icons);
+	void sendIcons(uint16_t icons);
 	void sendFYIBox(const std::string& message);
 
 	void sendDistanceShoot(const Position& from, const Position& to, uint8_t type);
